created containers folder & moved App component & its style sheet, also created components folder,ccreated StarRating component & its style sheet inside of it
