+++ src/components/StarRating/StarRating.jsx
@@ -1,0 +1,11 @@
+import "./StarRating.css"
+
+const starRating= ()=>{
+    return(
+        <div>
+          <h3>Rating</h3>
+        </div>
+    )
+}
+
+export default starRating;
