+++ src/contianers/App.js
@@ -1,10 +1,12 @@
 
 import './App.css';
+import StarRating from '../components/StarRating/StarRating';
 
 function App() {
   return (
     <div className="App">
       <h1>Star Rating</h1>
+      <StarRating />
     </div>
   );
 }
